fix: update-getting-started should wait for NuGet.org publishing
Added publish-nuget-org as dependency for update-getting-started job.

The GettingStarted repository references packages from NuGet.org, so it
should only be updated after packages are successfully published there.
Previously, it would try to update to a version that might not exist yet
on NuGet.org if the publishing failed or was still in progress.

The job still skips for pre-releases (if: !pre_release) as intended.

Signed-off-by: Kirill Kovalev <kirill.kovalev@abblix.com>